requirements.txt
astral==3.2
Flask==2.2.2
Flask==2.2.5
future==0.18.3
future==0.18.2
geopandas==0.14.0
influxdb_client==1.38.0
joblib==1.2.0
joblib==1.4.2
matplotlib==3.7.2
matplotlib==3.7.1
numba==0.57.0
numba==0.57.1
numba==0.58.1
numpy==1.23.5
numpy==1.24.3
pandas==2.0.3
PyKrige==1.7.1
pyproj==3.6.1
python-dotenv==1.1.0
pytz==2023.3.post1
pytz==2024.1
rasterio==1.4.1
scikit_learn==1.3.0
scipy==1.11.1
scipy==1.11.4
Shapely==2.1.1
SQLAlchemy==1.4.39
SQLAlchemy==2.0.37
tensorflow==2.18.0
tensorflow_intel==2.18.0
tqdm==4.65.0
tqdm==4.64.1
Werkzeug==2.2.2
Werkzeug==2.2.3
xarray==2023.6.0
~ensorflow_intel==2.18.0
